Create a conan package for prdemo
This should make it easier to access the archive contents from the
build system and developer machines.

This needs to be hooked up with CI/CD still.

There's lots of rough edges but any OS should be able to download the
package rather than needing to build from the recipe