reset_dungeon: Tool that resets dungeon
reset_dungeon.exe is a Win32 app that is shipped with Pre-ablo. When
run, it rerolls all the dungeons and resets all quest progress.

This can only run if the player is in town due to the complexity of the
same game loading system (that I haven't 100% decompiled...). This means
it can't be used to unbrick a game that's stuck. I'll look into that in
the future...